Registering AuthStore in constructor instead of StoreConfig
